+++ src/Pixie.Tests/ColorTests.cs
@@ -8,7 +8,7 @@ namespace Pixie.Tests
 
     public class ColorTests
     {
-        const float epsilon = 0.000001f;
+        const double epsilon = 0.000001;
 
         static readonly IEqualityComparer<Color> Comparer =
             Color.GetEqualityComparer(epsilon);
@@ -16,35 +16,35 @@ namespace Pixie.Tests
         [Fact]
         public void TestAddColors()
         {
-            var c1 = new Color(0.9f, 0.6f, 0.75f);
-            var c2 = new Color(0.7f, 0.1f, 0.25f);
-            var expected = new Color(1.6f, 0.7f, 1.00f);
+            var c1 = new Color(0.9, 0.6, 0.75);
+            var c2 = new Color(0.7, 0.1, 0.25);
+            var expected = new Color(1.6, 0.7, 1.00);
             Assert.Equal(expected, c1 + c2, Comparer);
         }
 
         [Fact]
         public void TestSubtractColors()
         {
-            var c1 = new Color(0.9f, 0.6f, 0.75f);
-            var c2 = new Color(0.7f, 0.1f, 0.25f);
-            var expected = new Color(0.2f, 0.5f, 0.5f);
+            var c1 = new Color(0.9, 0.6, 0.75);
+            var c2 = new Color(0.7, 0.1, 0.25);
+            var expected = new Color(0.2, 0.5, 0.5);
             Assert.Equal(expected, c1 - c2, Comparer);
         }
 
         [Fact]
         public void TestScaleColorByScalar()
         {
-            var c = new Color(0.2f, 0.3f, 0.4f);
-            var expected = new Color(0.4f, 0.6f, 0.8f);
+            var c = new Color(0.2, 0.3, 0.4);
+            var expected = new Color(0.4, 0.6, 0.8);
             Assert.Equal(expected, c * 2, Comparer);
         }
 
         [Fact]
         public void TestHadamardProduct()
         {
-            var c1 = new Color(1.0f, 0.2f, 0.4f);
-            var c2 = new Color(0.9f, 1.0f, 0.1f);
-            var expected = new Color(0.9f, 0.2f, 0.04f);
+            var c1 = new Color(1.0, 0.2, 0.4);
+            var c2 = new Color(0.9, 1.0, 0.1);
+            var expected = new Color(0.9, 0.2, 0.04);
             Assert.Equal(expected, c1 * c2, Comparer);
         }
     }
